+++ app/src/main/java/io/github/sds100/keymapper/UseCases.kt
@@ -57,6 +57,7 @@ object UseCases {
         ServiceLocator.accessibilityServiceAdapter(ctx),
         ServiceLocator.settingsRepository(ctx),
         ServiceLocator.purchasingManager(ctx),
+        ServiceLocator.ringtoneAdapter(ctx),
         getActionError(ctx),
         getConstraintError(ctx),
     )
@@ -71,6 +72,7 @@ object UseCases {
         ServiceLocator.cameraAdapter(ctx),
         ServiceLocator.soundsManager(ctx),
         ServiceLocator.shizukuAdapter(ctx),
+        ServiceLocator.ringtoneAdapter(ctx),
     )
 
     fun getConstraintError(ctx: Context) = GetConstraintErrorUseCaseImpl(
@@ -100,6 +102,7 @@ object UseCases {
     fun pauseKeyMaps(ctx: Context) = PauseKeyMapsUseCaseImpl(
         ServiceLocator.settingsRepository(ctx),
         ServiceLocator.mediaAdapter(ctx),
+        ServiceLocator.ringtoneAdapter(ctx),
     )
 
     fun showImePicker(ctx: Context): ShowInputMethodPickerUseCase = ShowInputMethodPickerUseCaseImpl(
@@ -163,6 +166,7 @@ object UseCases {
         ServiceLocator.soundsManager(ctx),
         ServiceLocator.permissionAdapter(ctx),
         ServiceLocator.notificationReceiverAdapter(ctx),
+        ServiceLocator.ringtoneAdapter(ctx),
     )
 
     fun detectKeyMaps(