+++ app/src/main/java/io/github/sds100/keymapper/actions/ActionDataEntityMapper.kt
@@ -1,5 +1,6 @@
 package io.github.sds100.keymapper.actions
 
+import androidx.core.net.toUri
 import io.github.sds100.keymapper.actions.pinchscreen.PinchScreenType
 import io.github.sds100.keymapper.actions.uielement.NodeInteractionType
 import io.github.sds100.keymapper.data.db.typeconverter.ConstantTypeConverters
@@ -239,11 +240,24 @@ object ActionDataEntityMapper {
             ActionId.PHONE_CALL -> ActionData.PhoneCall(number = entity.data)
 
             ActionId.SOUND -> {
-                val soundFileDescription =
-                    entity.extras.getData(ActionEntity.EXTRA_SOUND_FILE_DESCRIPTION)
-                        .valueOrNull() ?: return null
+                val isRingtoneUri = try {
+                    entity.data.toUri().scheme != null
+                } catch (e: Exception) {
+                    false
+                }
 
-                ActionData.Sound(soundUid = entity.data, soundDescription = soundFileDescription)
+                if (isRingtoneUri) {
+                    return ActionData.Sound.Ringtone(entity.data)
+                } else {
+                    val soundFileDescription =
+                        entity.extras.getData(ActionEntity.EXTRA_SOUND_FILE_DESCRIPTION)
+                            .valueOrNull() ?: return null
+
+                    ActionData.Sound.SoundFile(
+                        soundUid = entity.data,
+                        soundDescription = soundFileDescription,
+                    )
+                }
             }
 
             ActionId.VOLUME_INCREASE_STREAM,
@@ -659,7 +673,11 @@ object ActionDataEntityMapper {
         is ActionData.PinchScreen -> "${data.x},${data.y},${data.distance},${data.pinchType},${data.fingerCount},${data.duration}"
         is ActionData.Text -> data.text
         is ActionData.Url -> data.url
-        is ActionData.Sound -> data.soundUid
+        is ActionData.Sound -> when (data) {
+            is ActionData.Sound.Ringtone -> data.uri
+            is ActionData.Sound.SoundFile -> data.soundUid
+        }
+
         is ActionData.InteractUiElement -> data.description
         is ActionData.ControlMediaForApp.Rewind -> SYSTEM_ACTION_ID_MAP[data.id]!!
         is ActionData.ControlMediaForApp.Stop -> SYSTEM_ACTION_ID_MAP[data.id]!!
@@ -822,7 +840,7 @@ object ActionDataEntityMapper {
         is ActionData.Text -> emptyList()
         is ActionData.Url -> emptyList()
 
-        is ActionData.Sound -> listOf(
+        is ActionData.Sound.SoundFile -> listOf(
             EntityExtra(ActionEntity.EXTRA_SOUND_FILE_DESCRIPTION, data.soundDescription),
         )