+++ app/src/main/java/io/github/sds100/keymapper/home/HomeKeyMapListScreen.kt
@@ -17,6 +17,7 @@ import androidx.compose.foundation.layout.WindowInsetsSides
 import androidx.compose.foundation.layout.only
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.safeDrawing
+import androidx.compose.foundation.layout.systemBars
 import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.lazy.rememberLazyListState
@@ -172,7 +173,9 @@ fun HomeKeyMapListScreen(
         },
         listContent = {
             KeyMapList(
-                modifier = Modifier.animateContentSize(),
+                modifier = Modifier
+                    .windowInsetsPadding(WindowInsets.systemBars.only(sides = WindowInsetsSides.Horizontal))
+                    .animateContentSize(),
                 lazyListState = rememberLazyListState(),
                 listItems = state.listItems,
                 footerText = stringResource(R.string.home_key_map_list_footer_text),