+++ app/src/main/java/io/github/sds100/keymapper/system/accessibility/MyAccessibilityService.kt
@@ -22,9 +22,7 @@ import androidx.lifecycle.LifecycleRegistry
 import androidx.savedstate.SavedStateRegistry
 import androidx.savedstate.SavedStateRegistryController
 import androidx.savedstate.SavedStateRegistryOwner
-import io.github.sds100.keymapper.Constants
 import io.github.sds100.keymapper.R
-import io.github.sds100.keymapper.ServiceLocator
 import io.github.sds100.keymapper.actions.pinchscreen.PinchScreenType
 import io.github.sds100.keymapper.api.IKeyEventRelayServiceCallback
 import io.github.sds100.keymapper.api.KeyEventRelayService
@@ -40,7 +38,6 @@ import io.github.sds100.keymapper.util.InputEventType
 import io.github.sds100.keymapper.util.MathUtils
 import io.github.sds100.keymapper.util.Result
 import io.github.sds100.keymapper.util.Success
-import io.github.sds100.keymapper.util.onSuccess
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
@@ -202,14 +199,6 @@ class MyAccessibilityService :
 
     override fun onServiceConnected() {
         super.onServiceConnected()
-        val inputMethodAdapter = ServiceLocator.inputMethodAdapter(this)
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            inputMethodAdapter.getInfoByPackageName(Constants.PACKAGE_NAME).onSuccess {
-                softKeyboardController.setInputMethodEnabled(it.id, true)
-                softKeyboardController.switchToInputMethod(it.id)
-            }
-        }
 
         Timber.i("Accessibility service: onServiceConnected")
         lifecycleRegistry.currentState = Lifecycle.State.STARTED