+++ app/src/test/java/io/github/sds100/keymapper/actions/PerformActionsUseCaseTest.kt
@@ -86,6 +86,7 @@ class PerformActionsUseCaseTest {
             shizukuInputEventInjector = mock(),
             permissionAdapter = mock(),
             notificationReceiverAdapter = mock(),
+            ringtoneAdapter = mock(),
         )
     }